Update project to make use of Go Modules
Go modules where introduced in Golang 1.11 and is now the way how to manage
dependencies.

See https://golang.org/ref/mod for details.

This means, that the `vendor/` folder used by `Govendor` is obsolete and
replaced by the two files `go.mod` and `go.sum`, which have been added in this
commit.